saudapakka_backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
pgvector==0.2.4
django-cors-headers==4.3.1
gunicorn==21.2.0
Pillow==10.2.0
djangorestframework-simplejwt==5.3.1
django-filter==23.5
django-environ